TEN-87 - Some strukture and GET-Method for Game, User

+++ src/main/java/de/tenniskoenig/backend/domain/Game.java
@@ -1,5 +1,6 @@
 package de.tenniskoenig.backend.domain;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import javax.persistence.*;
 import java.io.Serializable;
 import java.sql.Time;
@@ -11,21 +12,21 @@ public class Game implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int gameID;
+    private long gameID;
     @Column(name = "player1Team1")
-    private int player1Team1;
+    private long player1Team1;
     @Column(name = "player2Team2")
-    private int player2Team2;
+    private long player2Team2;
     @Column(name = "player3Team1")
-    private int player3Team1;
+    private long player3Team1;
     @Column(name = "player4Team2")
-    private int player4Team2;
+    private long player4Team2;
     @Column(name = "results")
     private String results;
     @Column(name = "gameSetTeam1")
-    private int gameSetTeam1;
+    private long gameSetTeam1;
     @Column(name = "gameSetTeam2")
-    private int gameSetTeam2;
+    private long gameSetTeam2;
     @Column(name = "gameDate")
     private Date gameDate;
     @Column(name = "playTime")
@@ -55,7 +56,7 @@ public class Game implements Serializable {
         this.playTime = playTime;
     }
 
-    public int getGameID() {
+    public long getGameID() {
         return gameID;
     }
 
@@ -63,7 +64,7 @@ public class Game implements Serializable {
         this.gameID = gameID;
     }
 
-    public int getPlayer1Team1() {
+    public long getPlayer1Team1() {
         return player1Team1;
     }
 
@@ -71,7 +72,7 @@ public class Game implements Serializable {
         this.player1Team1 = player1Team1;
     }
 
-    public int getPlayer2Team2() {
+    public long getPlayer2Team2() {
         return player2Team2;
     }
 
@@ -79,7 +80,7 @@ public class Game implements Serializable {
         this.player2Team2 = player2Team2;
     }
 
-    public int getPlayer3Team1() {
+    public long getPlayer3Team1() {
         return player3Team1;
     }
 
@@ -87,7 +88,7 @@ public class Game implements Serializable {
         this.player3Team1 = player3Team1;
     }
 
-    public int getPlayer4Team2() {
+    public long getPlayer4Team2() {
         return player4Team2;
     }
 
@@ -103,7 +104,7 @@ public class Game implements Serializable {
         this.results = results;
     }
 
-    public int getGameSetTeam1() {
+    public long getGameSetTeam1() {
         return gameSetTeam1;
     }
 
@@ -111,7 +112,7 @@ public class Game implements Serializable {
         this.gameSetTeam1 = gameSetTeam1;
     }
 
-    public int getGameSetTeam2() {
+    public long getGameSetTeam2() {
         return gameSetTeam2;
     }
